+++ coolGuard-decision/src/main/java/cn/wnhyang/coolGuard/service/impl/SmsTemplateServiceImpl.java
@@ -5,10 +5,12 @@ import cn.hutool.core.util.StrUtil;
 import cn.wnhyang.coolGuard.context.FieldContext;
 import cn.wnhyang.coolGuard.convert.SmsTemplateConvert;
 import cn.wnhyang.coolGuard.entity.Action;
+import cn.wnhyang.coolGuard.entity.LabelValue;
 import cn.wnhyang.coolGuard.entity.SmsTemplate;
 import cn.wnhyang.coolGuard.mapper.SmsTemplateMapper;
 import cn.wnhyang.coolGuard.pojo.PageResult;
 import cn.wnhyang.coolGuard.service.SmsTemplateService;
+import cn.wnhyang.coolGuard.util.CollectionUtils;
 import cn.wnhyang.coolGuard.util.LFUtil;
 import cn.wnhyang.coolGuard.vo.create.SmsTemplateCreateVO;
 import cn.wnhyang.coolGuard.vo.page.SmsTemplatePageVO;
@@ -23,10 +25,12 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.regex.Pattern;
+
+import static cn.wnhyang.coolGuard.error.DecisionErrorCode.SMS_TEMPLATE_CODE_EXIST;
+import static cn.wnhyang.coolGuard.error.DecisionErrorCode.SMS_TEMPLATE_NOT_EXIST;
+import static cn.wnhyang.coolGuard.exception.util.ServiceExceptionUtil.exception;
 
 /**
  * 消息模版表 服务实现类
@@ -50,7 +54,11 @@ public class SmsTemplateServiceImpl implements SmsTemplateService {
     @Override
     @Transactional(rollbackFor = Exception.class)
     public Long create(SmsTemplateCreateVO createVO) {
+        if (smsTemplateMapper.selectByCode(createVO.getCode()) != null) {
+            throw exception(SMS_TEMPLATE_CODE_EXIST);
+        }
         SmsTemplate smsTemplate = SmsTemplateConvert.INSTANCE.convert(createVO);
+        smsTemplate.setParams(parseTemplateContentParams(smsTemplate.getContent()));
         smsTemplateMapper.insert(smsTemplate);
         return smsTemplate.getId();
     }
@@ -58,13 +66,23 @@ public class SmsTemplateServiceImpl implements SmsTemplateService {
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void update(SmsTemplateUpdateVO updateVO) {
-        SmsTemplate smsTemplate = SmsTemplateConvert.INSTANCE.convert(updateVO);
-        smsTemplateMapper.updateById(smsTemplate);
+        SmsTemplate smsTemplate = smsTemplateMapper.selectById(updateVO.getId());
+        if (smsTemplate == null) {
+            throw exception(SMS_TEMPLATE_NOT_EXIST);
+        }
+        SmsTemplate convert = SmsTemplateConvert.INSTANCE.convert(updateVO);
+        convert.setParams(parseTemplateContentParams(convert.getContent()));
+        smsTemplateMapper.updateById(convert);
     }
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void delete(Long id) {
+        SmsTemplate smsTemplate = smsTemplateMapper.selectById(id);
+        if (smsTemplate == null) {
+            throw exception(SMS_TEMPLATE_NOT_EXIST);
+        }
+        // TODO 引用
         smsTemplateMapper.deleteById(id);
     }
 
@@ -78,17 +96,18 @@ public class SmsTemplateServiceImpl implements SmsTemplateService {
         return smsTemplateMapper.selectPage(pageVO);
     }
 
+    @Override
+    public List<LabelValue> getLabelValueList() {
+        return CollectionUtils.convertList(smsTemplateMapper.selectList(), SmsTemplate::getLabelValue);
+    }
+
     @LiteflowMethod(value = LiteFlowMethodEnum.PROCESS, nodeId = LFUtil.SEND_SMS, nodeType = NodeTypeEnum.COMMON, nodeName = "加入名单组件")
     public void sendSms(NodeComponent bindCmp) {
         // TODO 完善
         Action.SendSms sendSms = bindCmp.getCmpData(Action.SendSms.class);
         SmsTemplate smsTemplate = smsTemplateMapper.selectByCode(sendSms.getSmsTemplateCode());
         FieldContext fieldContext = bindCmp.getContextBean(FieldContext.class);
-        Map<String, Object> map = new HashMap<>();
-        for (String param : smsTemplate.getParams()) {
-            map.put(param, fieldContext.getData2String(param));
-        }
-        String smsContent = StrUtil.format(smsTemplate.getContent(), map);
+        String smsContent = StrUtil.format(smsTemplate.getContent(), fieldContext);
         log.info("smsContent: {}", smsContent);
     }
